DW-706 build warning 제거.

+++ wdrbd9/send_buf.h
@@ -27,7 +27,7 @@ struct _send_req {
 struct ring_buffer {
 	char *name;
 	char *mem;
-	unsigned int length;
+	ULONGLONG length;
 	unsigned int read_pos;
 	unsigned int write_pos;
 	struct mutex cs;
@@ -53,9 +53,9 @@ struct _buffering_attr {
 
 typedef struct ring_buffer  ring_buffer;
 
-extern ring_buffer *create_ring_buffer(char *name, unsigned int length);
+extern ring_buffer *create_ring_buffer(char *name, ULONGLONG length);
 extern void destroy_ring_buffer(ring_buffer *ring);
-extern int get_ring_buffer_size(ring_buffer *ring);
+extern unsigned int get_ring_buffer_size(ring_buffer *ring);
 //extern void read_ring_buffer(ring_buffer *ring, char *data, int len);
 extern unsigned long read_ring_buffer(IN ring_buffer *ring, OUT char *data, OUT unsigned int* pLen);
 extern int write_ring_buffer(struct drbd_transport *transport, enum drbd_stream stream, ring_buffer *ring, const char *data, int len, int highwater, int retry);
